Remove broken dependency CI badge
I think dependencyci.com has been renamed to Tidelift and the badge URLs
have changed. See [this forum post] for more information. I am not
familiar with Dependencyci.com/Tidelift and/or what information it is
supposed to provide. I am not sure about the "example-base" being used
for the badge, either. I have removed it for the time being.

[this forum post]: https://forum.tidelift.com/t/new-urls-for-project-badges/288.